MAINT: Finish renaming utils
