+++ tests/fakeCookie.py
@@ -1,6 +1,6 @@
 import wailord.io as waio
 import wailord.exp as waex
-from wailord.utils import get_project_root
+from wailord._utils import get_project_root
 
 DATA_DIR = get_project_root() / "tests" / "data"
