+++ tests/fakeIO.py
@@ -1,5 +1,5 @@
 import wailord.io as waio
-from wailord.utils import get_project_root
+from wailord._utils import get_project_root
 
 DATA_DIR = get_project_root() / "tests" / "data"
