+++ tests/fakeInp.py
@@ -1,5 +1,5 @@
 import wailord.io as waio
-from wailord.utils import get_project_root
+from wailord._utils import get_project_root
 
 CONF_DIR = (
     get_project_root()
